Backend/requirements.txt
alembic==1.4.2
arrow==0.15.5
binaryornot==0.4.4
blinker==1.4
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cookiecutter==1.7.0
cryptography==2.8
dominate==2.5.1
Flask==1.1.1
Flask-Avatars==0.2.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
Flask-Dropzone==1.5.4
Flask-Ext==0.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-Script==2.0.6
Flask-Scss==0.5
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
future==0.18.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
jinja2-time==0.2.0
jwt==1.0.0
Mako==1.1.2
MarkupSafe==1.1.1
Pillow==7.1.2
poyo==0.5.0
pycparser==2.20
pyScss==1.3.7
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.12.1
python-socketio==4.5.1
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.8
visitor==0.1.3
Werkzeug==1.0.0
whichcraft==0.6.1
WTForms==2.2.1
